backend/requirements.txt
langchain
langchain-community
langchain-huggingface
huggingface-hub
python-dotenv
youtube-transcript-api
faiss-cpu
openai-whisper
yt-dlp
fastapi
uvicorn
torch
